yang: add breakout-reserved interface to cstore for Apollo NCP1-1 platform
Specify the breakout-reserved option to vyatta-interfaces.pl when
breaking out interfaces on the NCP1-1 platform to indicate which
interface has been reserved.

This allows the dataplane to take appropriate action, such as cleaning
up feature state to help out the FAL plugin. Whilst the mapping is
static on this platform, it allows the breakout-reserved interface to
be configured dynamically on platforms that support it.

Fixes: VRVDR-49850