vplane-route: use dataplane command for showing one MPLS route
When the user requests displaying a single MPLS route, rather than
getting all of the routes and filtering that information, which is
inefficient both for the dataplane and for this script, use a new
dataplane command to just retrieve the state of the one label.

An additional benefit to this is that now the dataplane can return
platform state associated with the label and this can be displayed
accordingly.

VRVDR-53194