vyatta-cfg :add support for enb/dis pause frame.
Add support for configuring (Set)Enable/Disable of
pause frame feature.
Also, add support for fetching(Get) the pause frame
configuration.
The changes are defining a yang model for the pause-frame
feature. It also includes the introduction of new script
and changes in the existing scripts for get and set
operations.

Feature: #VRVDR-49028