yang: correct transceiver data model to allow SFF-8472 vendor-rev
Fix the length of the vendor-rev to allow for SFF-8472 vendor
revisions, which are up to 4 characters longer, instead of up to 2
characters long.

Fix a comment in the vendor-oui field regarding the separator of the
OUI parts to conform with the dataplane implementation.

Fixes: VRVDR-50359
(cherry picked from commit 82c62e7cda45f065e37f05cc7e8fae58837bb57e)