dataplanestats: show link even if down
The dataplane will sometimes fail to report the speed associated with
links that are down. Generally, it depends on what the hardware decides
to report.

However, even if the link is down, show_interface_non_stats should report
what is knows about, such as the link status and the MTU.  Neither of
those are affected by link status.

Bugfix: VRVDR-51949