Kernel neighbor missing in show arp when there is no dataplane neighbor
While 'show arp <interface>' correctly displays kernel entries when
there is no matching dataplane entry, they are omitted for 'show arp'.
The reason is that the interface variable is set to an empty string
when not defined so as to form the command to retrieve the ARP entries.
However, the addition of kernel-only entries is conditional to the
interface not being defined for 'show arp'. A solution is to modify the
command directly, and not the interface variable.

VRVDR-52651

(cherry picked from commit d551d8624420c069a93226382c4bd67086d47b1a)