Behavior change in show arp output
With a matching dataplane entry, a kernel entry is shown as having the
same VALID or PENDING state. Without a matching dataplane entry, it is
incorrectly shown as VALID if it is in FAILED state. This is because it
is initially given an all-zeroes mac, and the entry is later deemed to
be VALID because the mac exists. A solution is to retain the state from
the kernel in this case.

VRVDR-47829

(cherry picked from commit d317350ad05120a300aaf989e6f473da83806b89)