Output of static entries in ARP table has changed
After the changes for VRVDR 47829, 'show arp' derives the controller
state from the kernel, rather than copying the dataplane state for an
ARP entry. If a MAC address is provided, an entry is considered to be
VALID without making a distinction for STATIC entries. A solution is to
map the kernel state so as to display VALID or STATIC.

VRVDR-52843

(cherry picked from commit 680bf057da8537a327555a30ae82ce32c6c7f6d3)