vyatta-op-transceiver: rename action_show_pause_info
The current call to get the pause info actually gets the
eth-info, the pause state being one member of that. Change the
routine that gets and prints the pause data to get the eth_info
and return it.  Use that data to get the pause value and print
that at the existing place.

VRVDR-46947