DataplaneStats: pass $sock to the dataplane if iterator cb funcs
There are several show commands that are callbacks of an iterator over
all relevant interfaces. Pass $sock as the 3rd arg into these callbacks
so that any that want to use it to help with formatting state have
access to it.  For those functions that have no need to use it is
ignored.

VRVDR-46947