Add the platform State to the dataplane vif interface output
The current command that shows a dataplane vif does not print out
the same style of output as when showig a dataplane interface. Instead
it prints out stats in the same format as 'show interface' using the
script in vyatta-interfaces.

Rather than change the current output  we want to append the
platform state info to the end of the current output so that we
don't break backwards compatibility.

Add and export a new function that will show the platform state for
a vif.  Modify the yang so that the script it calls gets passed an
extra argument, so that it can then query for the platform state.

VRVDR-46947