Add more efficient VIF VLAN validation
When the configd xpath plugin for VIF validation is installed we can validate VIFs at the interface level not the VIF level, which is much more efficient (one pass over 1000 VIFs versus 1000 passes!).  In the absence of the plugin, the must is essentially a no-op, with the original musts kicking in on the VIF node instead.

Part of VRVDR-53594.

Tested using existing acceptance tests.