Don't show the "Link detected" line output by ethtool
The "ethtool" command gets information from the kernel, which may not be
in-sync with the dataplane. The "Link detected:" line is one of these,
so remove it from the output displayed from "ethtool" when the op mode
command "show interfaces dataplane <if-name> physical" is used so as
not to confuse users.

VRVDR-49650