sfp: expand temperature and bias-current ranges
For transceiver-operating-temperature: if -32768 is returned,
then lower range value (dividing by 256) can be -128,
so change -127.99 to this value. Also, is 32767 is returned,
after rounding it would be 128, so change the upper range of 127.99
to this value.

For transceiver-laser-bias-current: if 65535 is returned, then
the upper range value (dividing by 500) can be 131.07, so change
131.00 to this value.

VRVDR-56385